requirements.txt
acres
annotated-types
anyio
APScheduler
asgiref
asttokens
attrs
azure-core
azure-devops
azure-functions
azure-identity
azure-storage-blob
azure-storage-file-share
beautifulsoup4
blinker
boto3
botocore
bs4
certifi
cffi
chardet
charset-normalizer
ci-info
click
colorama
comm
configobj
configparser
contourpy
cryptography
cycler
debugpy
decorator
distlib
Django
django-cors-headers
django-environ
djangorestframework
djangorestframework_simplejwt
easyocr
et_xmlfile
etelemetry
exceptiongroup
executing
fastapi
fastjsonschema
filelock
Flask
fonttools
fsspec
h11
httplib2
idna
imageio
imageio-ffmpeg
importlib_metadata
ipykernel
ipython
ipython_pygments_lexers
isodate
itsdangerous
jedi
Jinja2
jmespath
joblib
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
kiwisolver
lazy_loader
looseversion
lxml
MarkupSafe
matplotlib
matplotlib-inline
missingno
more-itertools
moviepy
mpmath
msal
msal-extensions
msrest
narwhals
nbformat
nest-asyncio
networkx
nibabel
ninja
nipype
numpy
oauthlib
opencv-python-headless
openpyxl
outcome
packaging
pandas
parso
pathlib
patsy
pillow
platformdirs
plotly
proglog
prompt_toolkit
prov
psutil
psycopg2-binary
pure_eval
puremagic
py4j
pyclipper
pycparser
pydantic
pydantic_core
pydot
Pygments
PyJWT
PyMuPDF
pyodbc
pyparsing
PyPDF2
PyPDF3
PySocks
pytesseract
python-bidi
python-dateutil
python-dotenv
pytz
pywin32
pyxnat
PyYAML
pyzmq
rdflib
referencing
requests
requests-file
requests-oauthlib
requests-toolbelt
rpds-py
s3transfer
scikit-image
scikit-learn
scipy
seaborn
selenium
setuptools
shapely
simple-salesforce
simplejson
six
sniffio
sortedcontainers
soupsieve
sqlparse
stack-data
starlette
statsmodels
sympy
threadpoolctl
tifffile
torch
torchvision
tornado
tqdm
traitlets
traits
trio
trio-websocket
typing_extensions
tzdata
tzlocal
urllib3
uvicorn
virtualenv
virtualenvwrapper-win
wcwidth
websocket-client
Werkzeug
wsproto
zeep
zipp
